requirements.txt
pymupdf
llama-cpp-python
sentence-transformers
faiss-cpu
numpy
scikit-learn
streamlit
fastapi
uvicorn
python-multipart
python-dotenv
pymongo
python-magic-bin